+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Exception handling associated with the reader allocation process.
+### Upgraded
+- "Keyple Service Resource Library" to version `2.1.1`
 
 ## [1.0.0] - 2023-04-25
 This is the initial release.